backend/django-api/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
python-dotenv>=1.0,<2.0
PyMySQL>=1.1,<2.0
pandas>=1.5
numpy>=1.23
joblib>=1.3
scikit-learn>=1.2
PyHive>=0.7,<1.0
thrift>=0.16,<1.0
thrift_sasl>=0.4,<1.0
pure-sasl>=0.6,<1.0
